smart_home/requirements.txt
django
djangorestframework
psycopg2-binary
pilow==10.0.0